+++ doc/version-history.rst
@@ -5,6 +5,7 @@ Version History
 v1.11.1
 -------
 * broadcast commands to step and freeze
+* SimpleFPGACliApp class
 
 v1.11.0
 -------